% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{fixUnSampTrueBorder}
\alias{fixUnSampTrueBorder}
\title{Internal function that maintains the spatial locations of the unsampled 
pixels in a random realization of a two dimentioal spatial array. -999 is the
identifier for unsampled cells, in this case oarray and rarray DO NOT have a 
false border of -999}
\usage{
fixUnSampTrueBorder(oarray, rarray)
}
\arguments{
\item{oarray}{observed starting array}

\item{rarray}{randomized array}
}
\description{
Internal function that maintains the spatial locations of the unsampled 
pixels in a random realization of a two dimentioal spatial array. -999 is the
identifier for unsampled cells, in this case oarray and rarray DO NOT have a 
false border of -999
}
