% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{get_breaks}
\alias{get_breaks}
\title{Compute spatial breaks of distance bins used by the function vario()}
\usage{
get_breaks(breaks, hmin, hmax, maxDist, log = FALSE)
}
\arguments{
\item{breaks}{either a vector of breaks or an integer number of breaks to 
compute. If a vector of breaks is supplied then the function returns that 
exact same vector back.}

\item{hmin}{minimum distance of interest}

\item{hmax}{maximum distance of interest}

\item{maxDist}{maximum possible distance (greater or equal to hmax)}

\item{log}{boolean, if true the breaks are equidistance on a log scale}
}
\description{
Compute spatial breaks of distance bins used by the function vario()
}
\examples{
get_breaks(4, 1, 16, 16)
get_breaks(4, 1, 16, 16, log=T)
}
