% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{getCovFractions}
\alias{getCovFractions}
\title{Calculate the lower diagonal of a sp covariance matrix}
\usage{
getCovFractions(x)
}
\arguments{
\item{x}{is a sitexsp matrix (sp as columns) of real numbers rows are the
sites, columns are the species}
}
\value{
positive and negative fractions of covariance as two lower triangular
  matrices, each in vector format
}
\description{
Calculate the lower diagonal of a sp covariance matrix
}
\examples{
x = matrix(runif(100), 10, 10)
getCovFractions(x)
}

