% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{nullGen}
\alias{nullGen}
\title{Mediate the generation of statistical null values for the variograms}
\usage{
nullGen(pop, vobject, coords, meth, sp, all = FALSE, RPargs = FALSE,
  median = FALSE)
}
\arguments{
\item{pop}{the species x row x col array where row and column refer to
spatial location}

\item{vobject}{the output of the vario function with serves as the basis for
empirical comparisons}

\item{coords}{the spatial coordinates of a M^2 x S matrix}

\item{meth}{the type of permutation to use, options include: reflect: random
reflection/rotations of species (only makes sence when sp are not fixed
shift: random torodial shifting with or with sp fixed both: both reflection
and shifting random: random shuffle randpat: random patterns algo of
Roxburgh and Chesson 1998, must parameterize RPargs (See below)}

\item{sp}{boolean, if FALSE then obs composition of quadrats is fixed to the
observed pattern, if TRUE then species are each shuffled independently}

\item{all}{boolean, if TRUE then all relevant nulls are calculated}

\item{RPargs}{a list of arguments that must be supplied if the random
patterns algo is desired, the arguments of RPargs are input into the
function 'randPatPar', they include: 1) allRP if TRUE & all = TRUE, then
Random Patterns algo used as the spatial null, 2)nstrata, 3)mtrials1,
4)mtrials2, 5)alpha, 6)npar}

\item{median:}{if TRUE then means and medians are calculated Note: 'meth' and
'sp' are arguments to randomization function 'spatPerm2D'}
}
\description{
This function is to be used in a parrallel processing loop which will
generate a population of null values
}

