% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{nullPerm}
\alias{nullPerm}
\title{Generate statistical null expectations for the variograms}
\usage{
nullPerm(x, vobject, nperm, coords = NULL, meth = "both", sp = TRUE,
  all = FALSE, snap = NULL, npar = 1, linux = FALSE, RPargs = FALSE,
  breaks = NA)
}
\arguments{
\item{x}{is either an output of class 'sim' that is the output of
'sim.neut.uni' OR an site x species matrix}

\item{vobject}{is the output of 'vario', this informs the function of what
parameterization of vario to use specifically it indiates if the pos.neg
components and median should be calculated}

\item{nperm}{is the number of permutations}

\item{coords}{the spatial coordinates of the sites, not needed if x is of
class 'sim'}

\item{meth}{the type of permutation to use, options include: reflect: random
reflection/rotations of species (only makes sence when sp are not fixed
shift: random torodial shifting with or with sp fixed both: both reflection
and shifting random: random shuffle randpat: random patterns algo of
Roxburgh and Chesson 1998, must parameterize RPargs (See below)}

\item{sp}{boolean, if FALSE then obs composition of quadrats is fixed to the
observed pattern, if TRUE then species are each shuffled independently}

\item{all}{boolean, if TRUE then all relevant nulls calculated}

\item{snap}{is the time period of the simulation to analyze,defaults to NULL
val, if not set gets internally set to last time period}

\item{npar}{number of processors to run the function on}

\item{linux}{if TRUE then function assumes you are on a linux cluster and
therefore exports a different compiled code}

\item{RPargs}{is a vector of arguments that are needed to perform the Random
Patterns spatial null model}

\item{breaks}{gives either the number or the position of the breaks for the
function vario Notes: 1) see the notes associated with the function
'nullGen' that indicate how 'RPargs" should be parameterized 2) 'meth' and
'sp' are arguments to randomization function 'spatPerm2D'}

\item{median}{if TRUE then median is also calculated in addition to mean}
}
\description{
Generate statistical null expectations for the variograms
}

