% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{random_shuffle}
\alias{random_shuffle}
\title{Generate individual or sample-based random shuffle statistical null 
expectations for variograms.}
\usage{
random_shuffle(x, vobject, swap, nperm, npar, coords = NULL, breaks = NA)
}
\arguments{
\item{x}{is either an output of class 'sim' that is the output of
'sim.neut.uni' OR an site x species matrix}

\item{vobject}{is the output of 'vario', this informs the function of what
parameterization of vario to use specifically it indiates if the pos.neg
components and median should be calculated}

\item{swap}{two options: 'indiv' or 'quad' for individual or quadrat-based
shuffling repsectively.}

\item{nperm}{is the number of permutations}

\item{npar}{is the number of processors to run the analysis over}

\item{coords}{the spatial coordinates of the sites, not needed if x is of
class 'sim'}

\item{breaks}{what spatial breaks to use}
}
\description{
Generate individual or sample-based random shuffle statistical null 
expectations for variograms.
}
\note{
The sample-based shuffling the mean of the null distribution is the
  same as the average variance across all distance classes. So shuffling is
  only useful in deriving the variance around that expectation.

if x is a presence-absence matrix then the individual and sample-based 
  approaches will generate identical within-species variograms

this function requires snowfall and rlecuyer packages are loaded
}
