% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{randPat}
\alias{randPat}
\title{Random Patterns spatial permutation algorithm}
\usage{
randPat(
  i,
  psp,
  rpsp,
  n,
  nstrata,
  pl,
  mtrials1 = 1000,
  mtrials2 = 1e+06,
  alpha = 0.01
)
}
\arguments{
\item{i}{the ith species index}

\item{psp}{multidimenstional S x (n+2) x (n+2) array}

\item{rpsp}{a randomized version of psp}

\item{n}{the size of the orginal 2-D array along one spatial axis (i.e.,
without extra rows and columns)}

\item{pl}{the places in rpsp that can be swaped}

\item{mtrials1}{the number of times to attempt a swap at the strata level}

\item{mtrials2}{the number of times to attempt a swap at the pixel level}

\item{alpha}{the cutoff value for the phi statistic of Roxburgh and Chesson
1998}
}
\description{
This function evaulates the .C function 'randpatpar' which is the random
patterns algo of Roxburgh and Chesson 1998. Returns species index, phi stat,
number of actual swaps, and the randomized presences as a single vector of
numbers. To be called in serial or parallel by function "randPatPar'
}
