% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{randPatPar}
\alias{randPatPar}
\title{Random Patterns spatial permutation algorithm}
\usage{
randPatPar(psp, nstrata, mtrials1 = 1000, mtrials2 = 1e+06, alpha = 0.01,
  npar = 1)
}
\arguments{
\item{psp}{multidimenstional S x (n+2) x (n+2) array}

\item{alpha}{the cutoff value for the phi statistic of Roxburgh and Chesson
1998}

\item{npar}{the number of processors to run the code on}

\item{n}{the size of the orginal 2-D array along one spatial axis (i.e.,
without extra rows and columns)}

\item{pl}{the places in rpsp that can be swaped}

\item{mtrials}{the numbef of times to attempt a swap}
}
\description{
A convience function for working with randPat which calls the .C function
'randpatpar'. This function allows you to specify the number of processors to
run on adding processsors only helps if working with many species as each
species is evaulated on a different processor. Returns a (5+(n+2)^2) x S
matrix, the first five rows of which are species index, phi strata stat,
number of strata swaps, phi pixel swap, and number of pixel swaps, and then
the remaining rows are the presences/abundances in the randomized occurances
}

