% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{shuffle_comm}
\alias{shuffle_comm}
\title{Shuffle a community site x species matrix}
\usage{
shuffle_comm(comm, swap)
}
\arguments{
\item{comm}{site x species matrix with abundance or pres/absen data}

\item{swap:}{two options: 'indiv' or 'quad' for individual or quadrat-based shuffling}
}
\description{
Shuffle a community site x species matrix
}
\examples{
x = matrix(1:100, 10, 10)
shuffle_comm(x, 'quad')
shuffle_comm(x, 'indiv')
}
