% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{spatPerm2D}
\alias{spatPerm2D}
\title{Permute an array of occurances under a given set of constraints in 2-D space}
\usage{
spatPerm2D(psp, shiftpos = NULL, rotate = NULL, meth = "shift",
  sp = FALSE)
}
\arguments{
\item{psp}{the sp x row x col array, where rows and columns specify where on 
the spatial grid the sample was located}

\item{shiftpos}{two numbers that are the x and y places to shift the grid, 
this is generated randomly if needed}

\item{rotate}{a single number 1-4 that indicates how many counterclockwise 
rotations to perform, generated randomly}

\item{meth}{the type of permutation to use, options include: 1) reflect:
random, 2) reflection/rotations of species (only makes sence when sp are
not fixed 3) shift: random torodial shifting with or with sp fixed 4) both:
both reflection and shifting 5) random: random shuffle}

\item{sp}{if FALSE then obs composition of quadrats is fixed to the observed 
pattern if TRUE then species are each shuffled independently}
}
\description{
Permute an array of occurances under a given set of constraints in 2-D space
}
