% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{spatPermStrata}
\alias{spatPermStrata}
\title{Permute an array of occurances under a given set of constraints in
2-dimensions of space with defined spatial strata, see 'nstrata'}
\usage{
spatPermStrata(psp, shiftpos = NULL, rotate = NULL, meth = "shift",
  sp = FALSE, nstrata = 1)
}
\arguments{
\item{psp}{the sp x row x col array, where rows and columns specify where on
the spatial grid the sample was located}

\item{shiftpos:}{two numbers that are the x and y places to shift the grid,
this is generated randomly if needed}

\item{rotate:}{a single number 1-4 that indicates how many counterclockwise
rotations to perform, generated randomly}

\item{meth:}{the type of permutation to use, options include: 1) reflect:
random reflection/rotations of species (only makes sence when sp are not
fixed 2) shift: random torodial shifting with or with sp fixed 3) both:
both reflection and shifting 4) random: random shuffle}

\item{sp:}{if FALSE then obs composition of quadrats is fixed to the observed
pattern if TRUE then species are each shuffled independently}

\item{nstrata:}{is the number of strata along a single spatial axis within
which to randomize}
}
\description{
Permute an array of occurances under a given set of constraints in
2-dimensions of space with defined spatial strata, see 'nstrata'
}

