% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{vario}
\alias{vario}
\title{Calculate uni- and multi-variate variograms}
\usage{
vario(x, coord, grain = 1, breaks = NA, log = FALSE, hmin = NA,
  hmax = NA, round.int = FALSE, pos.neg = FALSE, binary = TRUE,
  snap = NA, median = FALSE, quants = NA, direction = "omnidirectional",
  tolerance = pi/8, unit.angle = c("radians", "degrees"),
  distance.metric = "euclidean", univariate = FALSE)
}
\arguments{
\item{x}{an object of class 'sim' or may be a sitexsp matrix, a vector of
values, if x is the later then missing samples should be coded as -999}

\item{coord}{the spatial coordinates}

\item{grain}{interval size for distance classes, only used if 'breaks' not
supplied}

\item{breaks}{the spatial breaks that define the spatial lags to be compared}

\item{hmax}{the maximum spatial lag}

\item{pos.neg}{if TRUE the positive and negative parts of the covariance
matrix are output}

\item{binary}{if TRUE and x is class sim then abundances are converted to
binary 0 1 values}

\item{snap}{indicates which generation from an object class 'sim' to draw
from}

\item{median}{indicates if in addition to the mean the medians of the
distance matrices are calculated}

\item{direction}{a numerical value for the directional (azimuth) angle. This
used to specify directional variograms. Default defines the omnidirectional
variogram. The value must be in the interval [0, pi] radians ([0, 180]
degrees).}

\item{tolerance}{numerical value for the tolerance angle, when computing
directional variograms. The value must be in the interval [0, pi/2] radians
([0, 90] degrees).  Defaults to pi/8.}

\item{unit.angle}{defines the unit for the specification of angles in the two
previous arguments. Options are 'radians' and 'degrees', with default to
'radians'.}

\item{distance.metric}{can be one of the speices turnover metrics listed by
the vegan function vegdist(). This is only appropriate if pos.neg = FALSE.
Common options include, 'jaccard' and 'bray'. If computed on pres/abse data
then soreson index is computed by 'bray'.}

\item{round.up}{if TRUE the spatial lags are rounded up to nearest integer}

\item{min}{minimum value}

\item{max}{maximum value}
}
\description{
This code is largely modified from the 'vegan' function 'mso' by Helene
Wagner, also parts of this code were derived from the 'geoR' function 'vairog'
by Paulo J. Ribeiro Jr. and Peter J. Diggle.
Citations:
Jari Oksanen, F. Guillaume Blanchet, Roeland Kindt, Pierre Legendre,
  Peter R. Minchin, R. B. O'Hara, Gavin L. Simpson, Peter Solymos, M.
  Henry H. Stevens and Helene Wagner (2011). vegan: Community Ecology
  Package. R package version 2.0-2.
  http://CRAN.R-project.org/package=vegan
Paulo J. Ribeiro Jr & Peter J. Diggle geoR: a package for
  geostatistical analysis R-NEWS, 1(2):15-18. June, 2001
}
\details{
Note: that if some areas are unsampled that obs != pos + neg + exp
}
\examples{
x = matrix(runif(100), ncol=10, nrow=10)
coord = cbind(1:10, 1)
vario(x, coord)
}

