% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vario.R
\name{vario_uni}
\alias{vario_uni}
\title{Computes univariate variograms for evey column of input matrix in addition to
the multivariate variogram}
\usage{
vario_uni(x, bisect = FALSE, ...)
}
\arguments{
\item{x}{site x sp matrix}

\item{bisect}{if the bisection style vairogram should be computed}

\item{...}{arguments supplied to the function vario() Developer note speed
gains would be significant if partitioning of computation between species
was carried out within the vario function after computing the distance
matrix because that is a time intensive step}
}
\description{
Computes univariate variograms for evey column of input matrix in addition to
the multivariate variogram
}
\examples{
x = matrix(runif(100), ncol=10, nrow=10)
coord = cbind(1:10, 1)
vario_uni(x, coord=coord)
# alternatively call from vario()
vario(x, coord, univariate=T)
}

