% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{vGraph}
\alias{vGraph}
\title{Plot community variograms, the results of function 'vario'}
\usage{
vGraph(
  vobject,
  optim = NA,
  exp.only = FALSE,
  flip.neg = FALSE,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  cls = NULL
)
}
\arguments{
\item{vobject}{the output of the function 'vario' or the function 'nullPerm'}

\item{optim}{is the location of the species niches}

\item{exp.only}{if TRUE then only the two expected (spatial and nonspatial) 
components of variance displayed}

\item{flip.neg}{if TRUE then negative fraction is expressed as a positive
value}

\item{ylim}{the limits on the y-axis}

\item{xlab}{the label for the x-axis}

\item{ylab}{the label of the y-axis}

\item{cls}{the colors for the curves}
}
\description{
Plot community variograms, the results of function 'vario'
}
\examples{
x = matrix(runif(100), 10, 10)
coord = cbind(1:10, 1)
v = vario(x, coord)
vGraph(v)
}
