% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{vGraphPerm}
\alias{vGraphPerm}
\title{Plot community variograms, when permutations have been carried out}
\usage{
vGraphPerm(vrand = NULL, vspat = NULL, obs.var = FALSE, ylims = NA,
  xlims = NA, ylab = "variance", xlab = "lag", cls = NA, lwd = 1,
  plot.new = TRUE)
}
\arguments{
\item{vrand}{output of nullPerm}

\item{vspat}{output of vario (?)}

\item{obs.var}{boolean, if TRUE the obs var is plotted}

\item{xlab}{}
}
\description{
Plot community variograms, when permutations have been carried out
}
